+++ src/greynoise/__init__.py
@@ -1,7 +1,5 @@
 """GreyNoise API client and tools."""
 
-import logging
-
 from greynoise.api import GreyNoise  # noqa
 
 __author__ = "GreyNoise Intelligence"
@@ -11,6 +9,3 @@ __license__ = "MIT"
 __maintainer__ = "GreyNoise Intelligence"
 __email__ = "hello@greynoise.io"
 __status__ = "BETA"
-
-
-logging.getLogger(__name__).addHandler(logging.NullHandler())